+++ llm_mcp_server/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_mcp_handshake
+from . import test_mcp_oauth